requirements.txt
# SysMon - System Monitor Widget
# Cel Systems 2025

# Core dependencies
customtkinter>=5.2.0
psutil>=5.9.0
nvidia-ml-py>=12.535.0

# For CPU temperature on Windows
PyHardwareMonitor>=0.1.0

# System tray support
pystray>=0.19.0
Pillow>=10.0.0

# For transparency effects (optional)
pywin32>=306
